Disciplines and Projects extended.  - more templates  - previews  - project model improvements  - query optimizations

+++ resources/views/components/projects/preview.blade.php
@@ -1,13 +1,16 @@
+
 @pushonce('styles')
     <link rel="stylesheet" href="{{ asset('css/components/projects/preview.css') }}">
 @endpushonce
 
 
-<div class="project-preview">
+<div class="project-preview" style="background-image: url('{{asset('storage/'.$project->getPreviewPath())}}')">
     <a class="project-preview-link" href="gallery/{{ $project->id }}">
         <div class="project-actions">
-            <a class="project-preview-edit-link action-icon" title="Редактировать" href="/gallery/{{$project->id}}/edit"><img src="{{ asset('svg/Edit-project-white.svg') }}" alt="Редактировать"></a>
-            <a class="project-preview-delete-link action-icon" title="Удалить" href="/gallery/{{$project->id}}/delete"><img src="{{ asset('svg/Delete-project-white.svg') }}" alt="Удалить"></a>
+            @if($project->isAuthor(auth()->user()))
+                <a class="project-preview-edit-link action-icon" title="Редактировать" href="/gallery/{{$project->id}}/edit"><img src="{{ asset('svg/Edit-project-white.svg') }}" alt="Редактировать"></a>
+                <a class="project-preview-delete-link action-icon" title="Удалить" href="/gallery/{{$project->id}}/delete"><img src="{{ asset('svg/Delete-project-white.svg') }}" alt="Удалить"></a>
+            @endif
         </div>
         <div class="project-preview-bottom">
             <span class="project-title"> {{ $project->name }}</span>
